merge userProfiles into master (#51)
* Add codeCoverage

Add code coverage for a quick look over the code coverage.

* Added bage on readme.md

added Code coverage badge on the readme.md file, for a quick look of the code coverage.

* added omit of the coverage

added omit of the coverage, for a better sample of our coverage

* Refactoring of the coveragerc file

Refactoring of the coveragerc file

* Try resolve bug.

There is no more coverage since I added the coveragerc file. THis is a check to resolve the bug.

* Check if coveragerc is the problem

Check if coveragerc is the problem for the non-coverage

* Another way of excluding a part of code.

Another way of excluding a part of code, add pragma: no cover

* Added links for every user.

A user can now see every user profile page!

* Finishing seeing all viewwrs profile

A user can now see every users profiles.

* Refactoring of the navigation

refactoring of the navigation, this tackle issue #41

* Delete the user profiles view.

The users-profiles view is now not required. A user can follow another user by checking his page.

* Added dark red color to the like button 

Like button will be dark red if the user like this button

* User can now like a post on the profile view

User can like a post on the profile view.

* Changing color of following button

Button now change colors regarding if a user follow or not follow another user.

* Refactoring of the names

Refactoring of the title names.